Backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
sentence-transformers
httpx
pydantic